+++ pseudo_mixer/src/vado/temperature.h
@@ -14,25 +14,25 @@
 class Temperature
 {
   public:
-    Temperature(int temperaturePin, int betaValue, int seriesResistor = 10000, int thermistorNominal = 10000, int sampleCount = 5, int sampleWait = 1, int temperatureNominal = 25);
+    Temperature(uint8_t temperaturePin, uint16_t betaValue, uint16_t seriesResistor = 10000, uint16_t thermistorNominal = 10000, uint8_t sampleCount = 5, uint8_t sampleWait = 1, int8_t temperatureNominal = 25);
     const float GetCurrentTemp();
 
   private:
     // which analog pin to connect
-    int _temperaturePin = 0;
+    uint8_t _temperaturePin = 0;
     // resistance at 25 degrees C
-    int _thermistorNominal = 0;
+    uint16_t _thermistorNominal = 0;
     // temp. for nominal resistance (almost always 25 C)
-    int _temperatureNominal = 0;
+    int8_t _temperatureNominal = 0;
     // how many samples to take and average, more takes longer
     // but is more 'smooth'
-    int _sampleCount = 0;
+    uint8_t _sampleCount = 0;
     // how long to wait between samples
-    int _sampleWait = 0;
+    uint8_t _sampleWait = 0;
     // The beta coefficient of the thermistor (usually 3000-4000)
-    int _betaValue = 0;
+    uint16_t _betaValue = 0;
     // the value of the 'other' resistor
-    int _seriesResistor = 0;
+    uint16_t _seriesResistor = 0;
 };
 
 #endif
